9729043/UniMail/requirements.txt
Django>=4.0.6
psycopg2>=2.9.3
djangorestframework>=3.13.1